requirements.txt
absl-py==1.0.0
apted==1.0.3
astor==0.8.1
baidu-aip==2.2.18.0
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.12
dataclasses==0.8
gast==0.5.3
gensim==3.8.0
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.1.0
idna==3.3
importlib-metadata==4.8.3
jieba==0.42.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.6
numpy==1.16.4
opencv-python==3.4.0.14
pandas==0.20.3
protobuf==3.19.4
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
scikit-learn==0.20.3
scipy==1.5.2
six==1.16.0
smart-open==5.2.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
wincertstore==0.2
wrapt==1.13.3
zipp==3.6.0
